feat: hash refresh tokens
