+++ src/utils/index.ts
@@ -6,10 +6,11 @@ export * from './avatar';
 export * from './otp';
 export * from './gql-sdk';
 export * from './ticket';
-export * from './tokens';
+export * from './refresh-token';
 export * from './jwt';
 export * from './env';
 export * from './hasura';
 export * from './async-wrapper';
 export * from './webauthn';
 export * from './hasura-metadata';
+export * from './session';
