+++ src/utils/session.ts
@@ -1,45 +1,12 @@
 import { v4 as uuidv4 } from 'uuid';
-import { gqlSdk } from '@/utils';
-import { SignInResponse, Session } from '../types';
-import { UserFieldsFragment } from './__generated__/graphql-request';
+import { Session, SignInResponse } from '@/types';
+import { gqlSdk } from './gql-sdk';
 import { generateTicketExpiresAt } from './ticket';
 import { ENV } from './env';
-import { getUser } from './user';
 import { createHasuraAccessToken } from './jwt';
-
-function newRefreshExpiry() {
-  const date = new Date();
-
-  // cant return this becuase this will return a unix timestamp directly
-  date.setSeconds(date.getSeconds() + ENV.AUTH_REFRESH_TOKEN_EXPIRES_IN);
-
-  // instead we must return the js date object
-  return date;
-}
-
-const updateRefreshTokenExpiry = async (refreshToken: string) => {
-  await gqlSdk.getUsersByRefreshTokenAndUpdateRefreshTokenExpiresAt({
-    refreshToken,
-    expiresAt: new Date(newRefreshExpiry()),
-  });
-
-  return refreshToken;
-};
-
-export const getNewRefreshToken = async (
-  userId: string,
-  refreshToken = uuidv4()
-) => {
-  await gqlSdk.insertRefreshToken({
-    refreshToken: {
-      userId,
-      refreshToken,
-      expiresAt: new Date(newRefreshExpiry()),
-    },
-  });
-
-  return refreshToken;
-};
+import { getNewRefreshToken, updateRefreshTokenExpiry } from './refresh-token';
+import { getUser } from './user';
+import { UserFieldsFragment } from './__generated__/graphql-request';
 
 /**
  * Get new or update current user session
